+++ src/main/java/alcoholboot/toastit/feature/amazonimage/repository/ImageRepository.java
@@ -1,6 +1,6 @@
 package alcoholboot.toastit.feature.amazonimage.repository;
 
-import alcoholboot.toastit.global.domain.Image;
+import alcoholboot.toastit.feature.amazonimage.domain.Image;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
